created new folder for scripts, modified tsops_status.html to reflect changes
