+++ DashApp/templates/tsops_status.html
@@ -3,15 +3,10 @@
 
 {% block body %}
 
-
 <div class="container">
   <div class="row">
     <div class="col-sm-4">
-    <h1>Running Jobs</h1>
-        <div class="progress">
-            <div id="progress-bar" class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="" aria-valuemin="0" aria-valuemax="100" >
-            </div>
-        </div>
+    <h1>Running Jobs : {{ running_jobs }}</h1>
     </div>
   </div>
   
@@ -19,7 +14,7 @@
     <div class="col-md-8">
     <h1>Overall Service Completion ( 0 / {{ total_services }} )</h1>
         <div class="progress">
-            <div id="progress-bar" class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="" aria-valuemin="0" aria-valuemax="130" >
+            <div id="progress-bar" class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="" aria-valuemin="0" aria-valuemax={{ total_services}} >
             </div>
         </div>
     </div>
